# Performance issue due to inefficient copying/zeroing in STL