@ma30002000 2017-09-29
Travis build failed due to

    ./src/fuzzer/pkcs8.cpp:17:86:   required from here
    /usr/include/c++/4.8/bits/unique_ptr.h:65:22: error: invalid application of ‘sizeof’ to incomplete type 
    ‘Botan::Private_Key’
    static_assert(sizeof(_Tp)>0)

, which is unrelated to this request.

Also, travis build failes because linux libstdc++ 4.8 had not yet implemented std::is_trivially_constructible. This seems to have been added in libstdc++ 5.0.