@randombit 2017-09-29
Actually it appears in C++11 onwards we can skip defining `construct` and `destruct` entirely (see f.e. https://howardhinnant.github.io/allocator_boilerplate.html).

That may be a better option, assuming it works on all of our supported platforms.