@randombit 2017-09-29
It appears we can safely just remove `construct` and `destroy` entirely (#1229) which should resolve the issue, so I am closing this PR in favor of that one. It would be great if you can confirm this change resolves the issue for you. And thanks again for this profiling work.