@ma30002000 2017-10-04
Thanks for the quick fix. I can confirm that the performance issue on MSVC 2017 15.3.5 does not occur with 2.3.0.